+++ core/lib/multivm/src/versions/vm_1_3_2/vm_instance.rs
@@ -12,14 +12,13 @@ use zk_evm_1_3_3::{
 };
 use zksync_types::{
     l2_to_l1_log::{L2ToL1Log, UserL2ToL1Log},
-    tx::tx_execution_info::TxExecutionStatus,
     vm_trace::Call,
     L1BatchNumber, VmEvent, H256, U256,
 };
 
 use crate::{
     glue::GlueInto,
-    interface::{storage::WriteStorage, VmExecutionLogs},
+    interface::{storage::WriteStorage, TxExecutionStatus, VmExecutionLogs},
     versions::shared::{VmExecutionTrace, VmTrace},
     vm_1_3_2::{
         bootloader_state::BootloaderState,
