+++ core/node/state_keeper/src/updates/l2_block_updates.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use zksync_multivm::{
     interface::{
         CompressedBytecodeInfo, ExecutionResult, L2BlockEnv, TransactionExecutionResult,
-        VmExecutionResultAndLogs,
+        TxExecutionStatus, VmExecutionResultAndLogs,
     },
     vm_latest::TransactionVmExt,
 };
@@ -11,7 +11,7 @@ use zksync_types::{
     block::{BlockGasCount, L2BlockHasher},
     event::extract_bytecodes_marked_as_known,
     l2_to_l1_log::{SystemL2ToL1Log, UserL2ToL1Log},
-    tx::{tx_execution_info::TxExecutionStatus, ExecutionMetrics},
+    tx::ExecutionMetrics,
     vm_trace::Call,
     L2BlockNumber, ProtocolVersionId, StorageLogWithPreviousValue, Transaction, VmEvent, H256,
 };
